Added procedural placement based on mesh length
Uses component bounds and ray cast to get the position of mesh end. It Terquires testing on multiple mesh sizes. Had to reimport base meshes as XYZ coordinates were different to unreal and caused calculation erros. For example I would be looking for X length in UE but this would grab the X length from the asset which was a different vector